Refactor

+++ dotnet/Devices/Adafruit/LSM9DS1/Vector3.cs
@@ -2,23 +2,23 @@
 {
     public struct Vector3
     {
-        public float X;
-        public float Y;
-        public float Z;
+        public double X;
+        public double Y;
+        public double Z;
 
-        public Vector3(float x, float y, float z)
+        public Vector3(double x, double y, double z)
         {
             X = x;
             Y = y;
             Z = z;
         }
 
-        public static Vector3 operator *(Vector3 vector, float f)
+        public static Vector3 operator *(Vector3 vector, double f)
         {
             return new Vector3(vector.X * f, vector.Y * f, vector.Z * f);
         }
 
-        public static Vector3 operator /(Vector3 vector, float f)
+        public static Vector3 operator /(Vector3 vector, double f)
         {
             return new Vector3(vector.X / f, vector.Y / f, vector.Z / f);
         }
